Unify shader refl
