+++ axmol/rhi/vulkan/ShaderModuleVK.h
@@ -33,16 +33,6 @@ namespace ax::rhi::vk
  * @{
  */
 
-struct UniformBlockInfo
-{
-    int binding;          // Vulkan binding index
-    uint32_t sizeBytes;   // total size of the UBO
-    uint16_t numMembers;  // number of uniforms in this block
-    std::string name;     // block name
-};
-
-struct SLCReflectContext;
-
 /**
  * @brief A Vulkan-based ShaderModule implementation
  *
@@ -55,86 +45,9 @@ public:
 
     VkShaderModule internalHandle() const { return _shader; }
 
-    /**
-     * Get uniform info by engine built-in uniform enum name.
-     * @param name Specifies the engine built-in uniform enum name.
-     * @return The uniform location.
-     */
-    const UniformInfo& getUniformInfo(Uniform name) const;
-
-    /**
-     * Get uniform info by name.
-     * @param uniform Specifies the uniform name.
-     * @return The uniform location.
-     */
-    const UniformInfo& getUniformInfo(std::string_view name) const;
-
-    /**
-     * Get attribute location by engine built-in attribute enum name.
-     * @param name Specifies the engine built-in attribute enum name.
-     * @return The attribute location.
-     */
-    const VertexInputDesc* getVertexInputDesc(VertexInputKind name) const;
-
-    /**
-     * Get attribute location by attribute name.
-     * @param name Specifies the attribute name.
-     * @return The attribute location.
-     */
-    const VertexInputDesc* getVertexInputDesc(std::string_view name) const;
-
-    /**
-     * Get active attribute informations.
-     * @return Active attribute informations. key is attribute name and Value is corresponding attribute info.
-     */
-    inline const tlx::string_map<VertexInputDesc>& getActiveVertexInputs() const { return _activeVertexInputs; }
-
-    /**
-     * Get all uniformInfos.
-     * @return The uniformInfos.
-     */
-    inline const tlx::string_map<UniformInfo>& getActiveUniformInfos() const { return _activeUniformInfos; }
-
-    inline const std::vector<UniformBlockInfo>& getActiveUniformBlockInfos() const { return _activeUniformBlockInfos; }
-    inline const std::vector<UniformInfo>& getActiveSamplerInfos() const { return _activeSamplerInfos; }
-
-    /**
-     * Get maximum uniform location.
-     * @return Maximum uniform location.
-     */
-    inline const int getMaxLocation() const { return _maxLocation; }
-
-    /**
-     * Get uniform buffer size in bytes that holds all the uniforms.
-     * @return The uniform buffer size.
-     */
-    inline std::size_t getUniformBufferSize() const { return _uniformBufferSize; }
-
 private:
     void compileShader(VkDevice device, ShaderStage stage, std::string_view source);
-    void releaseShader();
-
-    void reflectVertexInputs(SLCReflectContext* context);
-    void reflectUniforms(SLCReflectContext* context);
-    void reflectSamplers(SLCReflectContext* context);
-
-    void setBuiltinLocations();
-
     VkShaderModule _shader = VK_NULL_HANDLE;
-
-    tlx::string_map<VertexInputDesc> _activeVertexInputs;
-    tlx::string_map<UniformInfo> _activeUniformInfos;
-
-    std::vector<UniformBlockInfo> _activeUniformBlockInfos;
-    std::vector<UniformInfo> _activeSamplerInfos;
-    int _samplerCount{0};
-
-    const VertexInputDesc* _builtinVertexInputs[VIK_COUNT];
-
-    int _maxLocation = -1;
-    UniformInfo _builtinUniforms[UNIFORM_COUNT];
-
-    std::size_t _uniformBufferSize = 0;
 };
 
 /** @} */
